requirements.txt
python-telegram-bot
requests 
Pillow 
img2pdf 
beautifulsoup4
python-dotenv
logging
asyncio